1153072 - do not delete /var/www/pulp_puppet on upgrade
Previously, the upgrade script would attempt to delete `/var/www/pulp_puppet`
as part of the upgrade procedure. This does not work since the upgrade runs as
apache user and not root.

Instead, move everything under `/var/www/pulp_puppet` to the correct location.
The directory itself is left for the user to clean up later as a manual step
since it requires root.