Updates metadata.json search to look in first top-level folder
The implementation was using a depth-first-search which did not
traverse in the same order on all platforms. Some Puppet modules
contain other modules as dependencies, and the DFS implementation
could incorrectly find the wrong one. This uses a very dumb
approach which looks for metadata.json in the top level folder
of the the tar.gz uploaded Puppet module. This is based on the
Puppet packaging guidelines [0].

[0]:  https://docs.puppetlabs.com/puppet/latest/reference/modules_fundamentals.html#module-fundamentals

https://pulp.plan.io/issues/890
closes #890